delete revisado

+++ src/main/java/com/grupo01/spring/repository/JuegoDao.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.grupo01.spring.model.Juego;
 
-public interface JuegoDao extends JpaRepository<Juego, Integer>{
+public interface JuegoDao extends JpaRepository<Juego, Long>{
